# internal/socket: reuse buffers in recv/sendMsgs