--- a/internal/socket/rawconn_msg.go
+++ b/internal/socket/rawconn_msg.go
@@ -55,7 +55,9 @@ func (c *Conn) sendMsg(m *Message, flags int) error {
 	vs := make([]iovec, len(m.Buffers))
 	var sa []byte
 	if m.Addr != nil {
-		sa = marshalInetAddr(m.Addr)
+		sa = make([]byte, sizeofSockaddrInet6)
+		n := marshalInetAddr(m.Addr, sa)
+		sa = sa[:n]
 	}
 	h.pack(vs, m.Buffers, m.OOB, sa)
 	var operr error